ai_fastapi_agent/ai-services/mcp_config.py
import os
from typing import Optional, Dict, Any
from dotenv import load_dotenv

load_dotenv()

class MCPConfig:
    def __init__(self):
        self.alloydb_engine = None
        self.cloudsql_engine = None
        self.is_configured = False
        
    def configure_alloydb(self, connection_string: str) -> bool:
        """Configure AlloyDB connection for MCP"""
        try:
            # For now, we'll use a simple in-memory store
            print("[MCP Config] Using in-memory store for development")
            self.is_configured = True
            return True
        except Exception as e:
            print(f"[MCP Config] Error configuring AlloyDB: {e}")
            return False
    
    def configure_cloudsql(self, connection_string: str) -> bool:
        """Configure CloudSQL connection for MCP"""
        try:
            # For now, we'll use a simple in-memory store
            print("[MCP Config] Using in-memory store for development")
            self.is_configured = True
            return True
        except Exception as e:
            print(f"[MCP Config] Error configuring CloudSQL: {e}")
            return False
    
    def get_memory_store(self) -> Optional[Any]:
        """Get the configured memory store"""
        # Return a simple dict for now
        return {}
    
    def is_ready(self) -> bool:
        """Check if MCP is properly configured"""
        return self.is_configured

# Global MCP configuration instance
mcp_config = MCPConfig()
