ai_fastapi_agent/ai-services/requirements.txt
# Core FastAPI and server dependencies
fastapi==0.115.12
uvicorn[standard]==0.30.6
gunicorn==23.0.0
python-dotenv==1.0.0
pydantic==2.10.4
python-multipart==0.0.9

# Google Cloud and Firebase dependencies
firebase-admin==6.7.0
google-cloud-firestore==2.20.1
google-generativeai==0.7.0
google-cloud-storage>=1.37.1,<3.0.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-api-python-client==2.166.0

# HTTP and async dependencies
requests==2.31.0
aiohttp==3.9.1

# Data processing and analysis
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0

# Email and notification services
jinja2==3.1.6
pyjwt==2.8.0

# Genetic analysis and file processing
PyPDF2==3.0.1
Pillow==10.2.0

# Security and authentication
cryptography==44.0.2

# Monitoring and logging
prometheus-client==0.19.0
structlog==24.1.0

# API documentation and validation
marshmallow==3.21.0
apispec==6.3.0

# Additional dependencies for new features
google-cloud-aiplatform==1.48.0

# LangChain dependencies for MCP
langchain>=0.2.0
langchain-google-genai==2.0.0
langchain-community>=0.2.0
